Moved Room component into module
