+++ src/tests/Room.test.js
@@ -2,4 +2,12 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import toJSON from 'enzyme-to-json';
 
-import {}
+import { rooms } from '../modules/Room';
+
+xdescribe('Room tests', () => {
+
+  it('each room has at least one door', () => {
+		const room = shallow(<Room )
+  });
+
+});
